+++ separados/q9.py
@@ -1,5 +1,5 @@
-peso = int(input('Digite seu peso em quilos: '))
-altura = int(input('Digite sua altura em centímetros: '))/100
+peso = float(input('Digite seu peso em quilos: '))
+altura = float(input('Digite sua altura em centímetros: '))/100
 imc = peso/(altura*altura)
 imc = round(imc, 2)
 if imc <= 18.5:
